+++ plom/cli/__init__.py
@@ -23,6 +23,9 @@ from .source_tools import upload_source
 from .source_tools import delete_source
 from .spec_tools import upload_spec
 from .task_tools import reset_task
+from .classlist_tools import delete_classlist
+from .classlist_tools import download_classlist
+from .classlist_tools import upload_classlist
 
 # what you get from "from plom.cli import *"
 __all__ = [
@@ -36,4 +39,7 @@ __all__ = [
     "upload_source",
     "upload_spec",
     "reset_task",
+    "delete_classlist",
+    "download_classlist",
+    "upload_classlist",
 ]