+++ plom_server/Preparation/services/classlist.py
@@ -3,6 +3,7 @@
 # Copyright (C) 2022 Edith Coates
 # Copyright (C) 2023 Natalie Balashov
 # Copyright (C) 2024-2025 Colin B. Macdonald
+# Copyright (C) 2025 Philip D. Loewen
 
 import csv
 import logging
@@ -64,17 +65,17 @@ class StagingStudentService:
         }
 
     def get_students_as_csv_string(self, *, prename: bool = False) -> str:
-        """Write the data from the classlist table into a string in CSV format.
+        """Write the classlist headers and data into a string in CSV format.
 
-        The header and name-columns are quoted.
+        Quote all headers and student names, but not ids or paper numbers.
         """
         txt = '"id","name","paper_number"\n'
         for row in self.get_students():
-            if prename and row["paper_number"]:
-                txt += f"{row['student_id']},\"{row['student_name']}\",{row['paper_number']}\n"
-            else:
-                # don't print the -1 for non-prename.
+            if not prename or row["paper_number"] is None or row["paper_number"] < 0:
+                # Leave paper_number empty when any of our non-prename sentinels appear.
                 txt += f"{row['student_id']},\"{row['student_name']}\",\n"
+            else:
+                txt += f"{row['student_id']},\"{row['student_name']}\",{row['paper_number']}\n"
         return txt
 
     @transaction.atomic()
@@ -130,11 +131,14 @@ class StagingStudentService:
         assert_can_modify_classlist()
         StagingStudent.objects.all().delete()
 
-    @transaction.atomic()
     def validate_and_use_classlist_csv(
         self, in_memory_csv_file: File, ignore_warnings: bool = False
     ) -> tuple[bool, list[dict[str, Any]]]:
-        """Validate and store the classlist from the in-memory file, if possible.
+        """Validate and store the classlist from the in-memory file, if possible, appending to existing classlist.
+
+        If there are no conflicts, this appends to an existing classlist.
+        The operation is atomic so either all new entries in the classlist
+        are added or none are.
 
         Args:
             in_memory_csv_file: some kind of Django file thing.
@@ -144,18 +148,19 @@ class StagingStudentService:
                 the validator expressed warnings.
 
         Returns:
-            2-tuple, the first entry is a bool indicating success.  In case of
-            success the 2nd entry is an empty list (or might contain
-            ignored warnings).  In case of errors the second list contains dicts
-            which elaborate on errors or warnings.
+            a 2-tuple (s,l), where ...
+            s is the boolean value of the statement "The operation succeeded",
+            l is a list of dicts describing warnings, errors, or notes.
+            When s is True, the list l may be empty or contain ignored warnings.
+            When s is False, the classlist in the database remains unchanged.
 
         Raises:
             PlomDependencyConflict: If dependencies not met.
         """
         assert_can_modify_classlist()
 
-        # now save the in-memory file to a tempfile and validate
-        # Note: we must be careful to unlink this file ourselves
+        # Save the in-memory file to a tempfile and validate it.
+        # Note: we must be careful to unlink this file ourselves.
         tmp_csv = Path(NamedTemporaryFile(delete=False).name)
 
         with open(tmp_csv, "wb") as fh:
@@ -172,13 +177,23 @@ class StagingStudentService:
             tmp_csv.unlink()
             return (success, werr)
 
-        # either no warnings, or warnings but ignore them - so read the csv
-        with open(tmp_csv) as fh:
-            csv_reader = csv.DictReader(fh, skipinitialspace=True)
+        werr = []
+
+        # Enforce empty-intersection between sets of incoming and known ID's.
+        known_ids = set([_["student_id"] for _ in self.get_students()])
+        known_paper_numbers = set(
+            [r.get("paper_number", -1) for r in self.get_students()]
+        )
+        new_ids = set()
+        new_paper_numbers = set()
+        # Note newline: https://docs.python.org/3/library/csv.html#id4
+        with open(tmp_csv, newline="") as fh:
+            prereader = csv.DictReader(fh)
+            headers = prereader.fieldnames
+            assert headers is not None  # Vlad would've noticed but MyPy doesn't know
+            # We accept "id", "ID", "Id", but code is messy #3822 #1140
+            # TODO: shouldn't this be Vlad's job?
             try:
-                # We accept "id", "ID", "Id", but code is messy #3822 #1140
-                headers = csv_reader.fieldnames
-                assert headers, "Expectedly empty csv header"
                 (id_key,) = [x for x in headers if x.casefold() == "id"]
                 (name_key,) = [x for x in headers if x.casefold() == "name"]
                 # paper_number is a bit harder b/c it might not be present
@@ -186,12 +201,70 @@ class StagingStudentService:
                 _tmp = [x for x in headers if x.casefold() == papernum_key]
                 if len(_tmp) == 1:
                     papernum_key = _tmp[0]
-                for row in csv_reader:
-                    self._add_student(
-                        row[id_key],
-                        row[name_key],
-                        paper_number=row.get(papernum_key, None),
-                    )
+            except ValueError as e:
+                success = False
+                errmsg = (
+                    "Bug in classlist validator?  Lower-level code did not notice "
+                    f"headers {headers} with repeats differing only in case: {str(e)}"
+                )
+                werr.append(
+                    {"warn_or_err": "error", "werr_line": None, "werr_text": errmsg}
+                )
+                tmp_csv.unlink()
+                return success, werr
+
+            for r in prereader:
+                new_ids.add(r[id_key])
+                # Next line correctly turns '' into -1:
+                new_paper_numbers.add(int(r.get(papernum_key, "-1") or "-1"))
+
+        known_paper_numbers.discard(-1)
+        new_paper_numbers.discard(-1)
+
+        id_overlap = known_ids & new_ids
+        paper_number_overlap = known_paper_numbers & new_paper_numbers
+
+        if False:
+            print("\nDEBUGGING classlist.py: Here are the known paper_numbers:")
+            print(known_paper_numbers)
+            print("DEBUGGING classlist.py: Here are the new paper_numbers:")
+            print(new_paper_numbers)
+            print("DEBUGGING classlist.py: Here is the set intersection:")
+            print(paper_number_overlap)
+            print(
+                f"DEBUGGING classlist.py: len(paper_number_overlap) = {len(paper_number_overlap)}."
+            )
+
+        if len(id_overlap) > 0:
+            success = False
+            errmsg = f"Incoming classlist collides with {len(id_overlap)} known ID's."
+            werr.append({"warn_or_err": "Error", "werr_text": errmsg})
+
+        if len(paper_number_overlap) > 0:
+            success = False
+            errmsg = f"Incoming classlist duplicates {len(paper_number_overlap)} paper numbers."
+            werr.append({"warn_or_err": "Error", "werr_text": errmsg})
+
+        if not success:
+            errmsg = "Server's classlist unchanged."
+            werr.append({"warn_or_err": "Warning", "werr_text": errmsg})
+            tmp_csv.unlink()
+            return success, werr
+
+        # Having developed some trust in the given CSV,
+        # we reopen it and transfer its contents into the server's classlist.
+        with open(tmp_csv, newline="") as fh:
+            csv_reader = csv.DictReader(fh, skipinitialspace=True)
+            try:
+                with transaction.atomic():
+                    # The 'atomic' wrapper gives the next loop an all-or-nothing property:
+                    # https://docs.djangoproject.com/en/5.2/topics/db/transactions/
+                    for row in csv_reader:
+                        self._add_student(
+                            row[id_key],
+                            row[name_key],
+                            paper_number=row.get(papernum_key, None),
+                        )
             except (IntegrityError, ValueError, KeyError, AssertionError) as e:
                 # in theory, we "asked permission" using vlad the validator
                 # so the input must be perfect and this can never fail---haha!